requirements.txt
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
pandas>=2.1.0
python-dotenv>=1.0.0
lxml>=4.9.0
aiofiles>=23.2.0
tenacity>=8.2.0
rich>=13.7.0
click>=8.1.0
html5lib>=1.1
